% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{dirichlet_laplace_logistic_mc}
\alias{dirichlet_laplace_logistic_mc}
\title{Logistic regression with Dirichlet Laplace prior}
\usage{
dirichlet_laplace_logistic_mc(Y, X, mc_length, a, beta_start)
}
\arguments{
\item{Y}{The binary response variable}

\item{X}{The design matrix}

\item{mc_length}{no. of MCMC iterations}

\item{beta_start}{initial parameters}
}
\value{
sample_beta: posterior samples of the coefficients beta

W_all: posterior samples of the latent variable

shi_all: posterior samples of the hyper-parameters

phi_all: posterior samples of the hyper-parameters

tau_sq_all: posterior samples of the hyper-parameters
}
\description{
Logistic regression with Dirichlet Laplace prior
}
