% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin.R
\name{dirichlet_laplace_negative_binomial_mc}
\alias{dirichlet_laplace_negative_binomial_mc}
\title{Negative Binomial regression with Dirichlet Laplace prior}
\usage{
dirichlet_laplace_negative_binomial_mc(
  Y,
  X,
  h,
  mc_length,
  a,
  starting_beta = NULL
)
}
\arguments{
\item{Y}{The count response}

\item{X}{The design matrix}

\item{h}{The number of failures in negative binomial regression}

\item{mc_length}{no. of MCMC samples}

\item{starting_beta}{initial parameters}
}
\value{
sample_beta: posterior samples of the coefficients beta

W_all: posterior samples of the latent variable

shi_all: posterior samples of the hyper-parameters

phi_all: posterior samples of the hyper-parameters

tau_sq_all: posterior samples of the hyper-parameters
}
\description{
Negative Binomial regression with Dirichlet Laplace prior
}
\examples{
dirichlet_laplace_negative_binomial_mc(Y=Y,X=X,h=h,mc_length=1000,a=0.8,starting_beta=NULL)
}
