% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb.R
\name{dirichlet_laplace_zinb_mc}
\alias{dirichlet_laplace_zinb_mc}
\title{Zero-inflated Negative Binomial Regression with Dirichlet Laplace prior}
\usage{
dirichlet_laplace_zinb_mc(
  X = X,
  y = Y,
  mc_length = mc_length,
  a = a,
  burnin = burnin,
  thin = thin
)
}
\arguments{
\item{X}{design matrix}

\item{y}{zero-inflated count response}

\item{mc_length}{no. of MCMC iterations}

\item{burnin}{burnin}

\item{thin}{thinning parameter, generally set to 1}
}
\value{
Alpha: posterior samples of Alpha parameter

Beta: posterior samples of Beta parameter

R or R2: posterior samples of R parameter

shi_all: posterior samples of hyper-parameter of the prior for Beta

fi_all:posterior samples of hyper-parameter of the prior for Beta

tau_sq_all:posterior samples of hyper-parameter of the prior for Beta

shi_all1: posterior samples of hyper-parameter of the prior for Alpha

fi_all1:posterior samples of hyper-parameter of the prior for Alpha

tau_sq_all1:posterior samples of hyper-parameter of the prior for Alpha
}
\description{
Zero-inflated Negative Binomial Regression with Dirichlet Laplace prior
}
