% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{double_pareto_logistic_mc}
\alias{double_pareto_logistic_mc}
\title{Logistic Regression with Double Pareto Prior}
\usage{
double_pareto_logistic_mc(Y, X, mc_length, beta_start)
}
\arguments{
\item{Y}{The binary response variable}

\item{X}{The design matrix}

\item{mc_length}{no. of MCMC iterations}

\item{beta_start}{initial parameters}
}
\value{
sample_beta: posterior samples of the coefficients beta

W_all: posterior samples of the latent variable

lambda_sq_all: posterior samples of the local parameter in the shrinkage prior

tau_sq_all: posterior samples of the global parameter in the shrinkage prior
}
\description{
Logistic Regression with Double Pareto Prior
}
