% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlogit.R
\name{double_pareto_mlogit_mc}
\alias{double_pareto_mlogit_mc}
\title{Multinomial logistic regression with Double Pareto prior}
\usage{
double_pareto_mlogit_mc(
  y,
  X,
  n = rep(1, nrow(as.matrix(y))),
  m.0 = array(0, dim = c(ncol(X), ncol(y))),
  sigma = array(0, dim = c(ncol(X), ncol(X), ncol(y))),
  samp = 1000,
  burn = 500
)
}
\arguments{
\item{y}{The categorical variable with more than two classes}

\item{X}{The design matrix}

\item{n}{sample size}

\item{m.0}{initial mean parameters mostly zero, a matrix of dimension P*J-1, p=no. of variables, J=no. of categorical classes}

\item{sigma}{initial variance covariance matrix of dimension P*P*J-1}

\item{samp}{no. of MCMC samples, generally 1000}

\item{burn}{burnin samples, generally 500}
}
\value{
sample_beta: posterior samples of the coefficients beta

w: posterior samples of the latent variable
}
\description{
Multinomial logistic regression with Double Pareto prior
}
