% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb.R
\name{double_pareto_zinb_mc}
\alias{double_pareto_zinb_mc}
\title{Zero-inflated Negative Binomial Regression with Double Pareto prior}
\usage{
double_pareto_zinb_mc(
  y = y,
  X = X,
  mc_length = mc_length,
  burnin = burnin,
  thin = 1
)
}
\arguments{
\item{y}{zero-inflated count response}

\item{X}{design matrix}

\item{mc_length}{no. of MCMC iterations}

\item{burnin}{burnin}

\item{thin}{thinning parameter, generally set to 1}
}
\value{
Alpha: posterior samples of Alpha parameter

Beta: posterior samples of Beta parameter

R or R2: posterior samples of R parameter

lambda_sq_all: posterior samples of hyper-parameter of the prior for Beta

tau_sq_all:posterior samples of hyper-parameter of the prior for Beta

lambda_sq_all1:posterior samples of hyper-parameter of the prior for Alpha

tau_sq_all1: posterior samples of hyper-parameter of the prior for Alpha
}
\description{
Zero-inflated Negative Binomial Regression with Double Pareto prior
}
