% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{horseshoe_logistic_mc}
\alias{horseshoe_logistic_mc}
\title{Logistic regression with Horseshoe prior}
\usage{
horseshoe_logistic_mc(Y, X, mc_length, beta_start)
}
\arguments{
\item{Y}{The binary response variable}

\item{X}{The design matrix}

\item{mc_length}{no. of MCMC iterations}

\item{beta_start}{initial parameters}
}
\value{
sample_beta: posterior samples of the coefficients beta

W_all: posterior samples of the latent variable

xi_all: posterior samples of hyper-parameter

gamma_all: posterior samples of hyper-parameter

lambda_all: posterior samples of local parameter in the shrinkage prior

tau_sq_all: posterior samples of global parameter in the shrinkage prior
}
\description{
Logistic regression with Horseshoe prior
}
