% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin.R
\name{horseshoe_negative_binomial_mc}
\alias{horseshoe_negative_binomial_mc}
\title{Negative Binomial regression with Horseshoe prior}
\usage{
horseshoe_negative_binomial_mc(Y, X, h, mc_length, starting_beta = NULL)
}
\arguments{
\item{Y}{The count response}

\item{X}{The design matrix}

\item{h}{The number of failures in negative binomial regression}

\item{mc_length}{no. of MCMC samples}

\item{starting_beta}{initial parameters
#' @export list of posterior samples of the parameters}
}
\value{
sample_beta: posterior samples of the coefficients beta

W_all: posterior samples of the latent variable

xi_all: posterior samples of hyper-parameter

gamma_all: posterior samples of hyper-parameter

lambda_all: posterior samples of local parameter in the shrinkage prior

tau_sq_all: posterior samples of global parameter in the shrinkage prior
}
\description{
Negative Binomial regression with Horseshoe prior
}
