% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb.R
\name{horseshoe_zinb_mc}
\alias{horseshoe_zinb_mc}
\title{Zero-inflated Negative Binomial Regression with Horseshoe prior}
\usage{
horseshoe_zinb_mc(
  X = X,
  y = Y,
  mc_length = mc_length,
  burnin = burnin,
  thin = thin
)
}
\arguments{
\item{X}{design matrix}

\item{y}{zero-inflated count response}

\item{mc_length}{no. of MCMC iterations}

\item{burnin}{burnin}

\item{thin}{thinning parameter, generally set to 1}
}
\value{
Alpha: posterior samples of Alpha parameter

Beta: posterior samples of Beta parameter

R or R2: posterior samples of R parameter

xi: posterior samples of hyper-parameter of the prior for Beta

gamma:posterior samples of hyper-parameter of the prior for Beta

lambda:posterior samples of hyper-parameter of the prior for Beta

tausq: posterior samples of hyper-parameter of the prior for Beta

xi1:posterior samples of hyper-parameter of the prior for Alpha

gamma1:posterior samples of hyper-parameter of the prior for Alpha

lambda1:posterior samples of hyper-parameter of the prior for Alpha

tausq1:posterior samples of hyper-parameter of the prior for Alpha
}
\description{
Zero-inflated Negative Binomial Regression with Horseshoe prior
}
